test: move tests and increase coverage
